Renamed MockSpanExporter to InMemorySpanExporter
