+++ exporters/CMakeLists.txt
@@ -2,8 +2,8 @@ if(WITH_OTPROTOCOL)
   add_subdirectory(otlp)
 endif()
 
-add_subdirectory(mock)
 add_subdirectory(ostream)
+add_subdirectory(memory)
 
 if(WITH_PROMETHEUS)
   add_subdirectory(prometheus)
