+++ exporters/memory/BUILD
@@ -1,12 +1,12 @@
 package(default_visibility = ["//visibility:public"])
 
 cc_library(
-    name = "mock_span_exporter",
+    name = "in_memory_span_exporter",
     srcs = [
-        "src/mock_span_exporter.cc",
+        "src/in_memory_span_exporter.cc",
     ],
     hdrs = [
-        "include/opentelemetry/exporters/mock/mock_span_exporter.h",
+        "include/opentelemetry/exporters/memory/in_memory_span_exporter.h",
     ],
     strip_include_prefix = "include",
     deps = [
