+++ exporters/memory/CMakeLists.txt
@@ -1,0 +1,3 @@
+include_directories(include)
+
+add_library(opentelemetry_exporter_in_memory_span src/in_memory_span_exporter.cc)
