--- exporters/mock/CMakeLists.txt
@@ -1,3 +1,0 @@
-include_directories(include)
-
-add_library(opentelemetry_exporter_mock_span src/mock_span_exporter.cc)
