+++ sdk/test/trace/simple_processor_test.cc
@@ -1,21 +1,21 @@
 #include "opentelemetry/sdk/trace/simple_processor.h"
-#include "opentelemetry/exporters/mock/mock_span_exporter.h"
+#include "opentelemetry/exporters/memory/in_memory_span_exporter.h"
 #include "opentelemetry/nostd/span.h"
 #include "opentelemetry/sdk/trace/span_data.h"
 
 #include <gtest/gtest.h>
 
 using namespace opentelemetry::sdk::trace;
-using opentelemetry::exporter::mock::MockSpanExporter;
+using opentelemetry::exporter::memory::InMemorySpanExporter;
 
-TEST(SimpleSpanProcessor, ToMockSpanExporter)
+TEST(SimpleSpanProcessor, ToInMemorySpanExporter)
 {
   std::shared_ptr<bool> shutdown_called(new bool(false));
 
   std::shared_ptr<std::vector<std::unique_ptr<SpanData>>> spans_received(
       new std::vector<std::unique_ptr<SpanData>>);
 
-  std::unique_ptr<SpanExporter> exporter(new MockSpanExporter(spans_received, shutdown_called));
+  std::unique_ptr<SpanExporter> exporter(new InMemorySpanExporter(spans_received, shutdown_called));
   SimpleSpanProcessor processor(std::move(exporter));
 
   auto recordable = processor.MakeRecordable();
